+++ IOS10-HW13-IsakovAnton/Sourses/Cells/SwitchCell.swift
@@ -1,0 +1,77 @@
+
+import UIKit
+
+class SwitchCell: UITableViewCell {
+    
+    // MARK: - Outlets
+    
+    private var isAirplaneModeEnabled = false
+    
+    let iconImageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.layer.cornerRadius = 4
+        imageView.clipsToBounds = true
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        return imageView
+    }()
+    
+    let titleLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    let switchControl: UISwitch = {
+        let switchControl = UISwitch()
+        switchControl.translatesAutoresizingMaskIntoConstraints = false
+        return switchControl
+    }()
+    
+    // MARK: - Initializers
+    
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        setupHierarchy()
+        setupLayout()
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func setupHierarchy() {
+        contentView.addSubview(iconImageView)
+        contentView.addSubview(titleLabel)
+        contentView.addSubview(switchControl)
+    }
+    
+    private func setupLayout() {
+        
+        NSLayoutConstraint.activate([
+            iconImageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
+            iconImageView.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
+            iconImageView.widthAnchor.constraint(equalToConstant: 24),
+            iconImageView.heightAnchor.constraint(equalToConstant: 24),
+            
+            titleLabel.leadingAnchor.constraint(equalTo: iconImageView.trailingAnchor, constant: 10),
+            titleLabel.trailingAnchor.constraint(equalTo: switchControl.leadingAnchor, constant: -10),
+            titleLabel.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
+            
+            switchControl.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
+            switchControl.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
+        ])
+    }
+    
+    func configuration (model: SwitchCells) {
+        iconImageView.image = model.image
+        iconImageView.backgroundColor = model.iconBackgroundColor
+        titleLabel.text = model.name
+        switchControl.isOn = isAirplaneModeEnabled
+        switchControl.addTarget(self, action: #selector(airplaneModeSwitchValueChanged(_:)), for: .valueChanged)
+        switchControl.isHidden = false
+    }
+    
+    @objc func airplaneModeSwitchValueChanged(_ sender: UISwitch) {
+        isAirplaneModeEnabled = sender.isOn
+    }
+}
